Update example, put pca-obs in a single file
